New cumul task time if same workstation

+++ class/ordre_fabrication_asset.class.php
@@ -2913,10 +2913,17 @@ class TAssetWorkstationOF extends TObjetStd{
         	$TIdOf = array($this->fk_assetOf);
         	$OF->getListeOFEnfants($PDOdb,$TIdOf);
         	krsort($TIdOf);
+            if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK)){
 
-        	$resIdTask = $db->query("SELECT MAX(t.rowid) as rowid
-            FROM ".MAIN_DB_PREFIX."projet_task t LEFT JOIN ".MAIN_DB_PREFIX."projet_task_extrafields tex ON (t.rowid=tex.fk_object)
-            WHERE t.fk_projet=".$OF->fk_project." AND tex.fk_of IN (".implode(',',$TIdOf).")");
+                $resIdTask = $db->query("SELECT MAX(t.rowid) as rowid
+                FROM " . MAIN_DB_PREFIX . "projet_task t LEFT JOIN " . MAIN_DB_PREFIX . "element_element ee  ON (ee.fk_target=t.rowid AND ee.targettype='project_task' AND ee.sourcetype='tassetof')
+                WHERE t.fk_projet=" . $OF->fk_project . " AND ee.fk_source IN (" . implode(',', $TIdOf) . ")");
+
+            }else {
+                $resIdTask = $db->query("SELECT MAX(t.rowid) as rowid
+                FROM " . MAIN_DB_PREFIX . "projet_task t LEFT JOIN " . MAIN_DB_PREFIX . "projet_task_extrafields tex ON (t.rowid=tex.fk_object)
+                WHERE t.fk_projet=" . $OF->fk_project . " AND tex.fk_of IN (" . implode(',', $TIdOf) . ")");
+            }
         	$objTask = $db->fetch_object($resIdTask);
         	$projectTask->fk_task_parent = (int)$objTask->rowid;
 
@@ -2954,6 +2961,11 @@ class TAssetWorkstationOF extends TObjetStd{
        	$projectTask->array_options['options_grid_use']=1;
        	$projectTask->array_options['options_fk_workstation']=$ws->getId();
 		$projectTask->array_options['options_fk_of']=$this->fk_assetOf;
+
+		$projectTask->add_object_linked('tassetof',$this->fk_assetOf);
+
+
+
 		$projectTask->date_c=dol_now();
 
 		$p = new Product($db);
@@ -2985,14 +2997,29 @@ class TAssetWorkstationOF extends TObjetStd{
 		$projectTask->fetch($this->fk_project_task);
 		$projectTask->fk_project = $OF->fk_project;
 
-		if($projectTask->planned_workload<=0)  $projectTask->planned_workload = $this->nb_hour*3600;
+        if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && !empty($OF->from_create)) {
+            $projectTask->planned_workload += $this->nb_hour * 3600;
+            $projectTask->add_object_linked('tassetof',$this->fk_assetOf);
+        } // On cumul le temps dans la tache
+        else if($projectTask->planned_workload <= 0) $projectTask->planned_workload = $this->nb_hour * 3600;
 
-		if(empty($conf->gantt->enabled)) {
-			$projectTask->date_start = strtotime(' +'.(int)$this->nb_days_before_beginning.'days',$OF->date_lancement);
-			$projectTask->date_end = $OF->date_besoin;
-			if($projectTask->date_end<$projectTask->date_start)$projectTask->date_end = $projectTask->date_start;
+        if(empty($conf->gantt->enabled)) {
+            if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && !empty($OF->from_create)) {
 
-		}
+                //On prend la date la plus petite
+                if($projectTask->date_start > $OF->date_lancement) $projectTask->date_start = strtotime(' +' . (int)$this->nb_days_before_beginning . 'days', $OF->date_lancement);
+
+                //On prend la date la plus grande
+                if($projectTask->date_end < $OF->date_besoin) $projectTask->date_end = $OF->date_besoin;
+
+                if($projectTask->date_end < $projectTask->date_start) $projectTask->date_end = $projectTask->date_start;
+            }
+            else {
+                $projectTask->date_start = strtotime(' +' . (int)$this->nb_days_before_beginning . 'days', $OF->date_lancement);
+                $projectTask->date_end = $OF->date_besoin;
+                if($projectTask->date_end < $projectTask->date_start) $projectTask->date_end = $projectTask->date_start;
+            }
+        }
 
 		$projectTask->update($user);
 
@@ -3042,7 +3069,27 @@ class TAssetWorkstationOF extends TObjetStd{
 
 		$action = '';
 
-		if ($of->fk_project > 0 && $this->fk_project_task == 0) $action = 'createTask';
+		if ($of->fk_project > 0 && $this->fk_project_task == 0){
+
+		    $action = 'createTask';
+
+            if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK)){
+
+                $taskstatic = new Task($db);
+                $TTask = $taskstatic->getTasksArray(null, null, $of->fk_project);
+                if(!empty($TTask)) {
+                    foreach($TTask as $task) {
+                        $task->fetch_optionals();
+                        if(!empty($task->array_options['options_fk_workstation']) && $this->fk_asset_workstation == $task->array_options['options_fk_workstation']){
+                            $action = 'updateTask';
+                            $this->fk_project_task=$task->id;
+                            $of->from_create=1;
+                        }
+
+                    }
+                }
+            }
+        }
 		elseif ($of->fk_project > 0 && $this->fk_project_task > 0) $action = 'updateTask';
 		elseif ($of->fk_project == 0 && $this->fk_project_task > 0) $action = 'deleteTask';
 
@@ -3055,7 +3102,7 @@ class TAssetWorkstationOF extends TObjetStd{
 				$this->updateTask($PDOdb, $db, $conf, $user, $of);
 				break;
 			case 'deleteTask':
-				$this->deleteTask($db, $conf, $user);
+                if(empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) || !empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && $this->isLastLink())$this->deleteTask($db, $conf, $user);
 				break;
 			default:
 				break;
@@ -3077,12 +3124,19 @@ class TAssetWorkstationOF extends TObjetStd{
 		else{
 			$db = &$this->db;
 		}
+        if(!empty($conf->global->ASSET_CUMULATE_PROJECT_TASK)){
+            $sql = "SELECT (SUM(tt.thm * tt.task_duration) / SUM(tt.task_duration)) as thm
+			FROM " . MAIN_DB_PREFIX . "projet_task_time tt
+			LEFT JOIN " . MAIN_DB_PREFIX . "projet_task_extrafields tex ON (tex.fk_object = tt.fk_task)
+			LEFT JOIN " . MAIN_DB_PREFIX . "element_element ee  ON (ee.fk_target=tt.fk_task AND ee.targettype='project_task' AND ee.sourcetype='tassetof')
+			WHERE ee.fk_source = " . $this->fk_assetOf . " AND tex.fk_workstation=" . $this->fk_asset_workstation . " AND tt.thm>0";
 
-		$sql="SELECT (SUM(tt.thm * tt.task_duration) / SUM(tt.task_duration)) as thm
-			FROM ".MAIN_DB_PREFIX."projet_task_time tt
-				LEFT JOIN ".MAIN_DB_PREFIX."projet_task_extrafields tex ON (tex.fk_object = tt.fk_task)
-			WHERE tex.fk_of = ".$this->fk_assetOf." AND tex.fk_workstation=".$this->fk_asset_workstation." AND tt.thm>0";
-
+        }else {
+            $sql = "SELECT (SUM(tt.thm * tt.task_duration) / SUM(tt.task_duration)) as thm
+			FROM " . MAIN_DB_PREFIX . "projet_task_time tt
+				LEFT JOIN " . MAIN_DB_PREFIX . "projet_task_extrafields tex ON (tex.fk_object = tt.fk_task)
+			WHERE tex.fk_of = " . $this->fk_assetOf . " AND tex.fk_workstation=" . $this->fk_asset_workstation . " AND tt.thm>0";
+        }
 		$res = $db->query($sql);
 		if($obj = $db->fetch_object($res)) {
 			if($obj->thm>0)	$this->thm = (float)$obj->thm;
@@ -3172,6 +3226,10 @@ class TAssetWorkstationOF extends TObjetStd{
 		$sql = 'DELETE FROM '.MAIN_DB_PREFIX.'element_element WHERE fk_source = '.(int) $this->rowid.' AND sourcetype = "tassetworkstationof" AND (targettype = "user" OR targettype = "task")';
 		$PDOdb->Execute($sql);
 
+		if( !empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && !empty($this->fk_assetOf)) {
+            $sql = 'DELETE FROM ' . MAIN_DB_PREFIX . 'element_element WHERE fk_source = ' . (int)$this->fk_assetOf . ' AND sourcetype = "tassetof" AND (targettype = "task")';
+            $PDOdb->Execute($sql);
+        }
 		if ($this->fk_project_task > 0)
 		{
 			require_once DOL_DOCUMENT_ROOT.'/projet/class/project.class.php';
@@ -3184,7 +3242,7 @@ class TAssetWorkstationOF extends TObjetStd{
 			if($projectTask->fetch($this->fk_project_task) > 0) {
 				// Suppression des occurences qui définissent cette tâches en tant que parente
 				$db->query('UPDATE '.MAIN_DB_PREFIX.'projet_task SET fk_task_parent = 0 WHERE fk_task_parent = '.$projectTask->id);
-				$projectTask->delete($user);
+                if(empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) || !empty($conf->global->ASSET_CUMULATE_PROJECT_TASK) && $this->isLastLink())$projectTask->delete($user);
 			}
 		}
 
@@ -3341,6 +3399,18 @@ class TAssetWorkstationOF extends TObjetStd{
 		$projectTask->progress = $progress;
 		$projectTask->update($db);
 	}
+
+	function isLastLink(){
+	    global $db;
+
+	    $sql = "SELECT rowid FROM ".MAIN_DB_PREFIX."asset_workstation_of WHERE fk_project_task=".$this->fk_project_task;
+	    $resql = $db->query($sql);
+	    $rows = $db->num_rows($resql);
+
+	    if($rows > 1) return false;
+
+	    return true;
+    }
 
 }
